--- aimmo-game/simulation/interactables/__init__.py
@@ -1,12 +1,0 @@
-from simulation.interactables.pickups import (ALL_PICKUPS, DamageBoostPickup,
-                                              HealthPickup,
-                                              InvulnerabilityPickup,
-                                              serialize_pickups)
-
-__all__ = [
-    'ALL_PICKUPS',
-    'DamageBoostPickup',
-    'InvulnerabilityPickup',
-    'HealthPickup',
-    'serialize_pickups'
-]
