+++ aimmo-game/simulation/interactables/pickups/__init__.py
@@ -1,0 +1,22 @@
+from simulation.interactables.pickups.damage_boost_pickup import \
+    DamageBoostPickup
+from simulation.interactables.pickups.health_pickup import HealthPickup
+from simulation.interactables.pickups.invulnerability_pickup import \
+    InvulnerabilityPickup
+
+
+def serialize_pickups(world_map):
+    return [cell.interactable.serialize() for cell in world_map.pickup_cells()]
+
+
+ALL_PICKUPS = (
+    DamageBoostPickup,
+    InvulnerabilityPickup,
+    HealthPickup,
+)
+__all__ = [
+    'ALL_PICKUPS',
+    'DamageBoostPickup',
+    'InvulnerabilityPickup',
+    'HealthPickup'
+]
