+++ aimmo-game/tests/functional/test_effect_expiry.py
@@ -2,7 +2,7 @@ import asyncio
 from unittest import TestCase
 
 from simulation.location import Location
-from simulation.interactables import DamageBoostPickup, InvulnerabilityPickup
+from simulation.interactables.pickups import DamageBoostPickup, InvulnerabilityPickup
 from simulation.interactables.effects import (DAMAGE_BOOST_DEFAULT,
                                               INVULNERABILITY_RESISTANCE)
