+++ aimmo-game/tests/functional/test_health_pickups_and_effects.py
@@ -5,10 +5,10 @@ from unittest import TestCase
 import hypothesis.strategies as st
 from hypothesis import given
 from simulation.location import Location
-from simulation.interactables import HealthPickup
+from simulation.interactables.pickups import HealthPickup
 from simulation.interactables.effects import (AVATAR_HEALTH_MAX,
-                                        HEALTH_RESTORE_DEFAULT,
-                                        HEALTH_RESTORE_MAX)
+                                              HEALTH_RESTORE_DEFAULT,
+                                              HEALTH_RESTORE_MAX)
 
 from .mock_world import MockWorld
 
@@ -38,9 +38,11 @@ class TestHealthPickupAndEffects(TestCase):
             }
         })
 
-        self.loop.run_until_complete(self.game.simulation_runner.run_single_turn(self.game.avatar_manager.get_player_id_to_serialized_action()))
+        self.loop.run_until_complete(self.game.simulation_runner.run_single_turn(
+            self.game.avatar_manager.get_player_id_to_serialized_action()))
 
-        self.assertEqual(self.cell.avatar, self.game.avatar_manager.get_avatar(1))
+        self.assertEqual(self.cell.avatar,
+                         self.game.avatar_manager.get_avatar(1))
         self.assertEqual(self.cell.avatar.health, self.initial_health +
                          HEALTH_RESTORE_DEFAULT)
 
@@ -55,7 +57,9 @@ class TestHealthPickupAndEffects(TestCase):
         avatar.health = 97
         self.cell.interactable = HealthPickup(self.cell)
 
-        self.loop.run_until_complete(self.game.simulation_runner.run_single_turn(self.game.avatar_manager.get_player_id_to_serialized_action()))
+        self.loop.run_until_complete(self.game.simulation_runner.run_single_turn(
+            self.game.avatar_manager.get_player_id_to_serialized_action()))
 
-        self.assertEqual(self.cell.avatar, self.game.avatar_manager.get_avatar(1))
+        self.assertEqual(self.cell.avatar,
+                         self.game.avatar_manager.get_avatar(1))
         self.assertEqual(self.cell.avatar.health, AVATAR_HEALTH_MAX)
