+++ aimmo-game/tests/functional/test_invulnerability_pickups_and_effects.py
@@ -2,7 +2,7 @@ import asyncio
 from unittest import TestCase
 
 from simulation.location import Location
-from simulation.interactables import InvulnerabilityPickup
+from simulation.interactables.pickups import InvulnerabilityPickup
 
 from .mock_world import MockWorld
