--- aimmo.code-workspace
@@ -1,20 +1,0 @@
-{
-	"folders": [
-		{
-			"path": "aimmo-game"
-		},
-		{
-			"path": "aimmo-game-worker"
-		},
-		{
-			"path": "aimmo-game-creator"
-		},
-		{
-			"path": "."
-		},
-		{
-			"path": "docs"
-		}
-	],
-	"settings": {}
-}
