+++ docs/architecture/games/interactable.md
